updated files and fixed code errors
